+++ backend/main.go
@@ -456,7 +456,7 @@ func main() {
 
 			privateServer.Use(util.NewTracer(util.PrivateGraph))
 			privateServer.Use(H.NewGraphqlTracer(string(util.PrivateGraph)).WithRequestFieldLogging())
-			privateServer.SetErrorPresenter(util.GraphQLErrorPresenter(string(util.PrivateGraph)))
+			privateServer.SetErrorPresenter(H.GraphQLErrorPresenter(string(util.PrivateGraph)))
 			privateServer.SetRecoverFunc(H.GraphQLRecoverFunc())
 			r.Handle("/",
 				privateServer,
@@ -500,7 +500,7 @@ func main() {
 				}))
 			publicServer.Use(util.NewTracer(util.PublicGraph))
 			publicServer.Use(H.NewGraphqlTracer(string(util.PublicGraph)))
-			publicServer.SetErrorPresenter(util.GraphQLErrorPresenter(string(util.PublicGraph)))
+			publicServer.SetErrorPresenter(H.GraphQLErrorPresenter(string(util.PublicGraph)))
 			publicServer.SetRecoverFunc(H.GraphQLRecoverFunc())
 			r.Handle("/",
 				publicServer,
